requirements.txt
flask
python-dotenv
requests
Adafruit_DHT
RPi.GPIO
